+++ src/quill_parser/src/lib.rs
@@ -1,6 +1,21 @@
+pub mod data_types;
+pub mod definition;
+pub mod expr_pat;
+pub mod file;
+pub mod identifier;
+pub mod types;
+pub mod visibility;
+
+use data_types::*;
+use definition::*;
+use expr_pat::*;
+use file::*;
+use identifier::*;
+use types::*;
+use visibility::*;
+
 use std::{
-    fmt::{Debug, Display},
-    hash::Hash,
+    fmt::Debug,
     iter::Peekable,
     ops::{Deref, DerefMut},
 };
@@ -39,6 +54,15 @@ impl DerefMut for TokenStream {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.iter
     }
+}
+
+/// Any top level item such as a definition or theorem.
+#[derive(Debug)]
+enum ItemP {
+    Use(UseP),
+    Data(DataP),
+    Enum(EnumP),
+    Definition(DefinitionP),
 }
 
 /// Parses a source file. This function parses the top-level item declarations, without parsing the contents of items.
@@ -149,73 +173,9 @@ impl<'input> Parser<'input> {
     }
 }
 
-/// A single `.ql` file may export data types and definitions.
-/// This `File` struct contains the parsed abstract syntax tree of a file.
-#[derive(Debug)]
-pub struct FileP {
-    pub uses: Vec<UseP>,
-    pub data: Vec<DataP>,
-    pub enums: Vec<EnumP>,
-    pub definitions: Vec<DefinitionP>,
-}
-
-/// Any top level item such as a definition or theorem.
-#[derive(Debug)]
-enum ItemP {
-    Use(UseP),
-    Data(DataP),
-    Enum(EnumP),
-    Definition(DefinitionP),
-}
-
-#[derive(Debug)]
-pub struct UseP {
-    pub source_file: IdentifierP,
-}
-
 ////////////////////
 //// DATA TYPES ////
 ////////////////////
-
-/// A `data` block, used to define product types.
-#[derive(Debug)]
-pub struct DataP {
-    pub vis: Visibility,
-    pub identifier: NameP,
-    pub type_params: Vec<TypeParameterP>,
-    pub type_ctor: TypeConstructorP,
-}
-
-/// Represents a type constructor in a `data` block.
-/// For example, `Just { value: T }`, where the `Just` is the `id`, and the `value` is the only element in `fields`.
-#[derive(Debug)]
-pub struct TypeConstructorP {
-    pub fields: Vec<FieldP>,
-}
-
-#[derive(Debug)]
-pub struct FieldP {
-    pub name: NameP,
-    pub ty: TypeP,
-}
-
-/// An `enum` block, used to define sum types.
-/// This kind of block allows you to associate arbitrary other types together into a single type.
-/// Any data type can be used inside an enum, even one used in a completely different enum.
-#[derive(Debug)]
-pub struct EnumP {
-    pub vis: Visibility,
-    pub identifier: NameP,
-    pub type_params: Vec<TypeParameterP>,
-    /// Has size 1 or larger.
-    pub alternatives: Vec<EnumVariantP>,
-}
-
-#[derive(Debug)]
-pub struct EnumVariantP {
-    pub name: NameP,
-    pub type_ctor: TypeConstructorP,
-}
 
 impl<'input> Parser<'input> {
     fn parse_use(&mut self) -> DiagnosticResult<UseP> {
@@ -355,54 +315,6 @@ impl<'input> Parser<'input> {
 //// DEFINITIONS AND EXPRESSIONS ////
 /////////////////////////////////////
 
-/// A `def` block. Defines a symbol's type and what values it takes under what circumstances.
-#[derive(Debug)]
-pub struct DefinitionP {
-    pub vis: Visibility,
-    pub name: NameP,
-    /// This definition might be defined with certain quantified type variables, e.g. foo[A, B].
-    pub type_parameters: Vec<TypeParameterP>,
-    pub definition_type: TypeP,
-    pub body: DefinitionBodyP,
-}
-
-#[derive(Debug)]
-pub enum DefinitionBodyP {
-    /// The body is defined as a series of cases to be pattern matched against.
-    PatternMatch(Vec<DefinitionCaseP>),
-    /// The body of the function is not written in Quill, it is an intrinsic and will be replaced
-    /// by hand-written LLVM IR code.
-    CompilerIntrinsic(Range),
-}
-
-/// Represents the loan conditions of a borrowed value.
-#[derive(Debug, Clone)]
-pub struct TypeBorrowP {
-    pub borrow_token: Range,
-    pub lifetime: LifetimeP,
-}
-
-/// Either the definition or use of a lifetime parameter.
-#[derive(Debug, Clone)]
-pub struct LifetimeP {
-    pub name: NameP,
-}
-
-#[derive(Debug, Clone)]
-pub struct TypeParameterP {
-    pub name: NameP,
-    /// A type variable may have one or more unnamed parameters, e.g. `F[_]` is a common type for a functor.
-    /// This field stores how many such parameters the type variable has.
-    pub parameters: u32,
-}
-
-/// Represents a case in a definition where we can replace the left hand side of a pattern with the right hand side.
-#[derive(Debug)]
-pub struct DefinitionCaseP {
-    pub pattern: ExprPatP,
-    pub replacement: ExprPatP,
-}
-
 impl<'input> Parser<'input> {
     /// `def ::= name named_type_params? ':' ty '{' def_body '}'
     fn parse_def(&mut self, vis: Visibility) -> DiagnosticResult<DefinitionP> {
@@ -492,36 +404,6 @@ impl<'input> Parser<'input> {
 ///////////////
 //// TYPES ////
 ///////////////
-
-#[derive(Debug)]
-pub enum TypeP {
-    /// An explicitly named type possibly with type parameters, e.g. `Bool`, `Either[T, U]` or `Foo[T[_]]`.
-    Named {
-        identifier: IdentifierP,
-        params: Vec<TypeP>,
-    },
-    /// A function `a -> b`.
-    /// Functions with more arguments, e.g. `a -> b -> c` are represented as
-    /// curried functions, e.g. `a -> (b -> c)`.
-    Function(Box<TypeP>, Box<TypeP>),
-    /// A borrowed type.
-    Borrow { ty: Box<TypeP>, borrow: TypeBorrowP },
-}
-
-impl Ranged for TypeP {
-    fn range(&self) -> Range {
-        match self {
-            TypeP::Named {
-                identifier,
-                params: args,
-            } => args
-                .iter()
-                .fold(identifier.range(), |acc, i| acc.union(i.range())),
-            TypeP::Function(left, right) => left.range().union(right.range()),
-            TypeP::Borrow { ty, borrow } => ty.range().union(borrow.borrow_token),
-        }
-    }
-}
 
 impl<'input> Parser<'input> {
     /// `type ::= type_borrow* (type_name type_args | "(" type ")") ("->" type)?`
@@ -696,113 +578,6 @@ impl<'input> Parser<'input> {
 //////////////////////////////////
 //// EXPRESSIONS AND PATTERNS ////
 //////////////////////////////////
-
-/// Represents either an expression or a pattern.
-#[derive(Debug)]
-pub enum ExprPatP {
-    /// A named variable e.g. `x` or `+`.
-    Variable(IdentifierP),
-    /// A primitive constant such as `14` or `false`.
-    Immediate { range: Range, value: ConstantValue },
-    /// Apply the left hand side to the right hand side, e.g. `a b`.
-    /// More complicated expressions e.g. `a b c d` can be desugared into `((a b) c) d`.
-    Apply(Box<ExprPatP>, Box<ExprPatP>),
-    /// A lambda abstraction, specifically `lambda params -> expr`.
-    Lambda {
-        lambda_token: Range,
-        params: Vec<NameP>,
-        expr: Box<ExprPatP>,
-    },
-    /// A let statement, specifically `let identifier = left_expr;`.
-    /// This kind of statement is only valid as an intermediate statement in a block.
-    Let {
-        let_token: Range,
-        name: NameP,
-        expr: Box<ExprPatP>,
-    },
-    /// A block of statements, inside parentheses, separated by newlines,
-    /// where the final expression in the block is the type of the block as a whole.
-    Block {
-        open_bracket: Range,
-        close_bracket: Range,
-        statements: Vec<ExprPatP>,
-    },
-    /// Borrow some data for a duration less than its full lifetime.
-    Borrow {
-        borrow_token: Range,
-        expr: Box<ExprPatP>,
-    },
-    /// Copy some data behind a borrow.
-    Copy {
-        copy_token: Range,
-        expr: Box<ExprPatP>,
-    },
-    /// The name of a data type, followed by brace brackets containing the data structure's fields.
-    ConstructData {
-        data_constructor: IdentifierP,
-        open_brace: Range,
-        close_brace: Range,
-        fields: ConstructDataFields,
-    },
-    /// An underscore `_` representing an unknown.
-    /// This is valid only in patterns, not normal expressions.
-    Unknown(Range),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum ConstantValue {
-    Unit,
-    Bool(bool),
-    Int(i64),
-}
-
-impl Display for ConstantValue {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ConstantValue::Unit => write!(f, "unit"),
-            ConstantValue::Bool(value) => write!(f, "bool {}", value),
-            ConstantValue::Int(value) => write!(f, "int {}", value),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct ConstructDataFields {
-    /// Fields that have been assigned values, e.g. `foo = 1`.
-    pub fields: Vec<(NameP, ExprPatP)>,
-    /// Fields that have not been assigned values (so will inherit their value from the local variable with that name), e.g. `foo`.
-    /// This is useful in patterns, where fields are often not assigned different names.
-    pub auto_fields: Vec<NameP>,
-}
-
-impl Ranged for ExprPatP {
-    fn range(&self) -> Range {
-        match self {
-            ExprPatP::Variable(identifier) => identifier.range(),
-            ExprPatP::Immediate { range, .. } => *range,
-            ExprPatP::Apply(left, right) => left.range().union(right.range()),
-            ExprPatP::Unknown(range) => *range,
-            ExprPatP::Lambda {
-                lambda_token, expr, ..
-            } => lambda_token.union(expr.range()),
-            ExprPatP::Let {
-                let_token, expr, ..
-            } => let_token.union(expr.range()),
-            ExprPatP::Block {
-                open_bracket,
-                close_bracket,
-                ..
-            } => open_bracket.union(*close_bracket),
-            ExprPatP::Borrow { borrow_token, expr } => borrow_token.union(expr.range()),
-            ExprPatP::Copy { copy_token, expr } => copy_token.union(expr.range()),
-            ExprPatP::ConstructData {
-                data_constructor,
-                close_brace,
-                ..
-            } => data_constructor.range().union(close_brace.range()),
-        }
-    }
-}
 
 /// An internal structure used when parsing expression blocks.
 struct ExprBlockBody {
@@ -1304,140 +1079,6 @@ fn collapse_func_application(terms: Vec<(Option<Operator>, ExprPatP)>) -> Functi
 /////////////////////////
 //// OTHER UTILITIES ////
 /////////////////////////
-
-/// An unresolved identifier, which is a string of text at some range in code.
-#[derive(Debug, Clone)]
-pub struct IdentifierP {
-    /// Must contain at least one segment.
-    pub segments: Vec<NameP>,
-}
-
-#[derive(Debug, Clone)]
-struct Operator {
-    level: u32,
-    name: NameP,
-    ty: AssociativityType,
-}
-
-#[derive(Debug, Clone, Copy)]
-enum AssociativityType {
-    NonAssociative,
-    InfixR,
-    InfixL,
-}
-
-impl IdentifierP {
-    /// If this identifier represents a name that has a level of associativity,
-    /// Returns the properties of the given operator.
-    fn as_operator(&self) -> Option<Operator> {
-        if let Some(name) = self.segments.last() {
-            as_operator_inner(name.clone())
-        } else {
-            None
-        }
-    }
-}
-
-/// Returns the properties of the given operator.
-fn as_operator_inner(name: NameP) -> Option<Operator> {
-    let n = name.name.as_str();
-    if name.name.as_str().chars().next().unwrap().is_alphanumeric() {
-        None
-    } else {
-        Some(if n.contains(':') {
-            Operator {
-                level: 5,
-                name,
-                ty: AssociativityType::InfixR,
-            }
-        } else if n.contains('+') || n.contains('-') {
-            Operator {
-                level: 10,
-                name,
-                ty: AssociativityType::InfixL,
-            }
-        } else if n.contains('*') || n.contains('/') {
-            Operator {
-                level: 15,
-                name,
-                ty: AssociativityType::InfixL,
-            }
-        } else if n.contains('=') || n.contains('<') || n.contains('>') {
-            Operator {
-                level: 3,
-                name,
-                ty: AssociativityType::NonAssociative,
-            }
-        } else {
-            Operator {
-                level: 1,
-                name,
-                ty: AssociativityType::InfixL,
-            }
-        })
-    }
-}
-
-impl Ranged for IdentifierP {
-    fn range(&self) -> Range {
-        self.segments
-            .iter()
-            .fold(self.segments[0].range, |range, segment| {
-                range.union(segment.range)
-            })
-    }
-}
-
-impl Display for IdentifierP {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        for (i, segment) in self.segments.iter().enumerate() {
-            if i != 0 {
-                write!(f, "::")?;
-            }
-            write!(f, "{}", segment)?;
-        }
-        Ok(())
-    }
-}
-
-/// A name for an item, which cannot be qualified.
-#[derive(Debug, Clone, Eq)]
-pub struct NameP {
-    pub name: String,
-    pub range: Range,
-}
-
-impl PartialEq for NameP {
-    fn eq(&self, other: &Self) -> bool {
-        self.name == other.name
-    }
-}
-
-impl Hash for NameP {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.name.hash(state)
-    }
-}
-
-impl Display for NameP {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.name)
-    }
-}
-
-pub enum Visibility {
-    Public(Range),
-    Private,
-}
-
-impl Debug for Visibility {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Visibility::Public(_) => write!(f, "public"),
-            Visibility::Private => write!(f, "private"),
-        }
-    }
-}
 
 impl<'input> Parser<'input> {
     /// Parses an identifier if the next token is an identifier fragment.
