+++ src/quill_parser/src/visibility.rs
@@ -1,0 +1,17 @@
+use std::fmt::Debug;
+
+use quill_common::location::Range;
+
+pub enum Visibility {
+    Public(Range),
+    Private,
+}
+
+impl Debug for Visibility {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Visibility::Public(_) => write!(f, "public"),
+            Visibility::Private => write!(f, "private"),
+        }
+    }
+}
